requirements.txt
requests
python-dotenv
google-api-python-client
google-auth
playwright
